deleted phewas references in settings.py
