+++ settings.py
@@ -73,11 +73,6 @@ class Config:
     CRISPR_FILENAME2 = file_or_resource('crispr_descriptions.tsv')
     CRISPR_EVIDENCE_FILENAME = 'crispr-21-08-2019.json'
 
-    # PheWAS catalog
-    PHEWAS_CATALOG_FILENAME = file_or_resource('phewas-catalog-19-10-2018.csv')
-    PHEWAS_CATALOG_EVIDENCE_FILENAME = f"phewas_catalog-{datetime.today().strftime('%Y-%m-%d')}.json"
-    PHEWAS_CATALOG_W_CONSEQUENCES = 'https://storage.googleapis.com/otar000-evidence_input/PheWAS/data_files/phewas_w_consequences.csv'
-
     # Gene2Phenotype
     #G2P_FILENAME = 'DDG2P.csv.gz'
     G2P_DD_FILENAME = file_or_resource('DDG2P_2_4_2020.csv.gz')
